requirements.txt
pytorch==1.1.0
numpy==1.16.3
